refactor: dto이름 변경에 따른 후속 조치
